feat!: extract hooks from component files
