+++ src/AccordionButton.tsx
@@ -2,15 +2,11 @@ import * as React from 'react';
 import { useContext } from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
-import AccordionContext, {
-  isAccordionItemSelected,
-  type AccordionEventKey,
-} from './AccordionContext';
+import AccordionContext, { isAccordionItemSelected } from './AccordionContext';
 import AccordionItemContext from './AccordionItemContext';
 import type { BsPrefixProps, BsPrefixRefForwardingComponent } from './helpers';
 import { useBootstrapPrefix } from './ThemeProvider';
-
-type EventHandler = React.EventHandler<React.SyntheticEvent>;
+import useAccordionButton from './useAccordionButton';
 
 export interface AccordionButtonProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement>,
@@ -26,37 +22,6 @@ const propTypes = {
   /** A callback function for when this component is clicked */
   onClick: PropTypes.func,
 };
-
-export function useAccordionButton(
-  eventKey: string,
-  onClick?: EventHandler,
-): EventHandler {
-  const { activeEventKey, onSelect, alwaysOpen } = useContext(AccordionContext);
-
-  return (e) => {
-    /*
-      Compare the event key in context with the given event key.
-      If they are the same, then collapse the component.
-    */
-    let eventKeyPassed: AccordionEventKey =
-      eventKey === activeEventKey ? null : eventKey;
-    if (alwaysOpen) {
-      if (Array.isArray(activeEventKey)) {
-        if (activeEventKey.includes(eventKey)) {
-          eventKeyPassed = activeEventKey.filter((k) => k !== eventKey);
-        } else {
-          eventKeyPassed = [...activeEventKey, eventKey];
-        }
-      } else {
-        // activeEventKey is undefined.
-        eventKeyPassed = [eventKey];
-      }
-    }
-
-    onSelect?.(eventKeyPassed, e);
-    onClick?.(e);
-  };
-}
 
 const AccordionButton: BsPrefixRefForwardingComponent<
   'div',
