+++ src/useCol.ts
@@ -1,0 +1,61 @@
+import classNames from 'classnames';
+import {
+  useBootstrapPrefix,
+  useBootstrapBreakpoints,
+  useBootstrapMinBreakpoint,
+} from './ThemeProvider';
+import type { ColProps, ColOrder, ColSize, NumberAttr } from './Col';
+
+export interface UseColMetadata {
+  as?: React.ElementType;
+  bsPrefix: string;
+  spans: string[];
+}
+
+export default function useCol({
+  as,
+  bsPrefix,
+  className,
+  ...props
+}: ColProps): [any, UseColMetadata] {
+  bsPrefix = useBootstrapPrefix(bsPrefix, 'col');
+  const breakpoints = useBootstrapBreakpoints();
+  const minBreakpoint = useBootstrapMinBreakpoint();
+
+  const spans: string[] = [];
+  const classes: string[] = [];
+
+  breakpoints.forEach((brkPoint) => {
+    const propValue = props[brkPoint];
+    delete props[brkPoint];
+
+    let span: ColSize | undefined;
+    let offset: NumberAttr | undefined;
+    let order: ColOrder | undefined;
+
+    if (typeof propValue === 'object' && propValue != null) {
+      ({ span, offset, order } = propValue);
+    } else {
+      span = propValue;
+    }
+
+    const infix = brkPoint !== minBreakpoint ? `-${brkPoint}` : '';
+
+    if (span)
+      spans.push(
+        span === true ? `${bsPrefix}${infix}` : `${bsPrefix}${infix}-${span}`,
+      );
+
+    if (order != null) classes.push(`order${infix}-${order}`);
+    if (offset != null) classes.push(`offset${infix}-${offset}`);
+  });
+
+  return [
+    { ...props, className: classNames(className, ...spans, ...classes) },
+    {
+      as,
+      bsPrefix,
+      spans,
+    },
+  ];
+}
