+++ www/docs/components/accordion.mdx
@@ -95,7 +95,7 @@ this can be achieved with a custom toggle that is context aware and also takes a
 ### useAccordionButton
 
 ```jsx
-import { useAccordionButton } from 'react-bootstrap/AccordionButton';
+import useAccordionButton from 'react-bootstrap/useAccordionButton';
 
 const decoratedOnClick = useAccordionButton(eventKey, onClick);
 ```
